changed the format of returning data
