+++ server/controllers/updateBooking.js
@@ -2,37 +2,8 @@ import {
 	checkAvailabilityForUpdate,
 	getBookingById,
 	updateBookingQuery,
-	
 } from "./dataAccess.js";
-import { isPast } from "./validators.js";
-
-export const formatDateTime = (isoDateString) => {
-	const date = new Date(isoDateString);
-	const formattedDateTime = new Intl.DateTimeFormat("en-US", {
-		year: "numeric",
-		month: "short",
-		day: "2-digit",
-		hour: "numeric",
-		minute: "numeric",
-		second: "numeric",
-		hour12: false,
-		timeZone: "Europe/London",
-	}).format(date);
-
-	return formattedDateTime;
-};
-
-export const formatTime = (isoDateString) => {
-	const date = new Date(isoDateString);
-	const options = {
-		hour: "2-digit",
-		minute: "2-digit",
-		second: "2-digit",
-		timeZone: "Europe/London",
-	};
-
-	return new Intl.DateTimeFormat("en-GB", options).format(date);
-};
+import { isPast, formatDateTime, formatTime } from "./validators.js";
 
 export const updateBooking = async (req, res) => {
 	try {
