+++ server/controllers/validators.js
@@ -16,10 +16,10 @@ export const validateBooking = async (booking, errors) => {
 
 	if (!userId) {
 		errors.push(new ErrorMessage("User not found"));
-		return
+		return;
 	}
 
-	const formattedDate = new Date(`${date}T12:30:00.000Z`);
+	const formattedDate = formatDateTime(`${date}T12:30:00.000Z`);
 
 	if (isPast(formattedDate)) {
 		errors.push(new ErrorMessage("Please enter a valid date"));
@@ -48,4 +48,33 @@ export const validateBooking = async (booking, errors) => {
  */
 export const isPast = (date) => {
 	return isNaN(date) || date < new Date();
+};
+
+// format date and time
+export const formatDateTime = (isoDateString) => {
+	const date = new Date(isoDateString);
+	const formattedDateTime = new Intl.DateTimeFormat("en-US", {
+		year: "numeric",
+		month: "short",
+		day: "2-digit",
+		hour: "numeric",
+		minute: "numeric",
+		second: "numeric",
+		hour12: false,
+		timeZone: "Europe/London",
+	}).format(date);
+
+	return formattedDateTime;
+};
+
+export const formatTime = (isoDateString) => {
+	const date = new Date(isoDateString);
+	const options = {
+		hour: "2-digit",
+		minute: "2-digit",
+		second: "2-digit",
+		timeZone: "Europe/London",
+	};
+
+	return new Intl.DateTimeFormat("en-GB", options).format(date);
 };
